% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hfun_g_t}
\alias{hfun_g_t}
\title{Single trait logistic curve fit}
\usage{
hfun_g_t(par, Times)
}
\arguments{
\item{par}{numeric of curve par}

\item{Times}{numeric of times}
}
\value{
vector of logistic curve number
}
\description{
Single trait logistic curve fit
}
