% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funmap.R
\name{hfun_get_ini}
\alias{hfun_get_ini}
\title{hfun_get_ini Initialization of fitting mean and covariance matrix parameters}
\usage{
hfun_get_ini(
  phe,
  Times,
  ranks,
  control = list(cur_lower = rep(0.001, 3), cur_upper = rep(1000, 3), cur_control =
    list(itermax = 500, trace = F), cov_lower = 0.001, cov_upper = 100, cov_control =
    list(itermax = 500, trace = F))
)
}
\arguments{
\item{phe}{matrix:input phenotype data}

\item{Times}{numeric of times like c(1,2,3,6,14,19,25,48)}

\item{ranks}{numerc the number of traits}

\item{control}{list include
cur_lower        numeric of curve parameters lower set, default is rep(1e-3,3)
cur_upper        numeric of curve parameters upper set, default is rep(1000,3)
cur_control      list of curve DEoptim control, default is list(itermax=500,trace=F), more details see ?DEoptim
cov_lower        numeric of cov-matrix parameters lower set, default is 1e-3
cov_upper        numeric of cov-matrix parameters upper set, default is 100
cov_control      list of cov-matrix DEoptim control, default is list(itermax=500,trace=F), more details see ?DEoptim}
}
\value{
numeric parameters of curve and cov-matrix
}
\description{
hfun_get_ini Initialization of fitting mean and covariance matrix parameters
}
