% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hfun_get_invsigma}
\alias{hfun_get_invsigma}
\title{Multi-trait SAD1 inv & det}
\usage{
hfun_get_invsigma(par, Times, ranks)
}
\arguments{
\item{par}{numeric of covariance matrix par}

\item{Times}{numeric of times}

\item{ranks}{int, number of traits}
}
\value{
list $sigma_1=sigma_1;$det=det_sigma_1
}
\description{
Multi-trait SAD1 inv & det
}
