% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hfun_get_mu}
\alias{hfun_get_mu}
\title{Multi-trait logistic curve fit}
\usage{
hfun_get_mu(par, Times)
}
\arguments{
\item{par}{numeric of curve par}

\item{Times}{numeric of times}
}
\value{
matrix dim = 1\emph{length_n.trait}length_n.times
}
\description{
Multi-trait logistic curve fit
}
