% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hfun_H0_mle}
\alias{hfun_H0_mle}
\title{H0 calculate mle}
\usage{
hfun_H0_mle(par, N_phe, pheT, ranks)
}
\arguments{
\item{par}{numeric of mle par}

\item{N_phe}{input phenotype matrix}

\item{pheT}{numeric of times}

\item{ranks}{int, number of traits}
}
\value{
number of mle value
}
\description{
H0 calculate mle
}
