% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hfun_H1_mle}
\alias{hfun_H1_mle}
\title{H1 calculate mle}
\usage{
hfun_H1_mle(par, pheno, Times, ranks, ng)
}
\arguments{
\item{par}{numeric of mle par}

\item{pheno}{list of pheno}

\item{Times}{numeric of times}

\item{ranks}{int, number of traits}

\item{ng}{int, number of geno type}
}
\value{
number of mle value
}
\description{
H1 calculate mle
}
