% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funmap.R
\name{hfun_iden_fit}
\alias{hfun_iden_fit}
\title{hfun_iden_fit individual curve fit}
\usage{
hfun_iden_fit(x, Times, ranks)
}
\arguments{
\item{x}{vector of individual pheno}

\item{Times}{numeric of times like c(1,2,3,6,14,19,25,48)}

\item{ranks}{numerc the number of traits}
}
\value{
vector, which is composed of ranks part of curve_par a,b,r and R^2
}
\description{
hfun_iden_fit individual curve fit
}
