% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funmap.R
\name{hfun_load_data}
\alias{hfun_load_data}
\title{hfun_load_data    Initial database load}
\usage{
hfun_load_data(
  address.p,
  address.g,
  Times,
  ranks = NULL,
  missing = FALSE,
  log.p = TRUE,
  fit.check = FALSE
)
}
\arguments{
\item{address.p}{charactor like "/home/C/data/pheno.csv"}

\item{address.g}{charactor like "/home/C/data/SNP.csv",missing value is numeric -9}

\item{Times}{numeric of times like c(1,2,3,6,14,19,25,48)}

\item{ranks}{numerc the number of traits}

\item{missing}{logical whether geno database have missing value, default is False}

\item{log.p}{logical whether transform pheno to log10.pheno, default is True}

\item{fit.check}{logical whether do individual curve fit; return mean R^2}
}
\value{
list dat.p,dat.g,dat.t,dat.r,dat.m
}
\description{
hfun_load_data    Initial database load
}
