% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funmap.R
\name{hfun_lr_cal}
\alias{hfun_lr_cal}
\title{hfun_lr_cal LR calculate function}
\usage{
hfun_lr_cal(
  format_data,
  interval,
  useinv = TRUE,
  usecpp = TRUE,
  ini_control = NULL,
  LR_only = FALSE
)
}
\arguments{
\item{format_data}{data from hfun_load_data}

\item{interval}{numeric the serial number of genotype you want to calculate eg.c(1:10) or c(1,5,9,12,56)}

\item{useinv}{logical whether use inverse matrix, default is TRUE}

\item{usecpp}{logical whether use Rcpp code, default is TRUE}

\item{ini_control}{list,default is NULL, include
cur_lower        numeric of curve parameters lower set, default is rep(1e-3,3)
cur_upper        numeric of curve parameters upper set, default is rep(1000,3)
cur_control      list of curve DEoptim control, default is list(itermax=500,trace=F), more details see ?DEoptim
cov_lower        numeric of cov-matrix parameters lower set, default is 1e-3
cov_upper        numeric of cov-matrix parameters upper set, default is 100
cov_control      list of cov-matrix DEoptim control, default is list(itermax=500,trace=F), more details see ?DEoptim}

\item{LR_only}{logical whether output only include LR value, default is FALSE}
}
\value{
matrix of lr_calculate form example c(LR,H1_mle$value,H1_mle$par,H0_mle$value,H0_mle$par)
}
\description{
hfun_lr_cal LR calculate function
}
