% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funmap.R
\name{hfun_permutation}
\alias{hfun_permutation}
\title{hfun_permutation do permutation test}
\usage{
hfun_permutation(format_data, permu_times = 5, interval, cutp = 0.05)
}
\arguments{
\item{format_data}{data from hfun_load_data}

\item{permu_times}{int, times you want to do permutation test,default is 5}

\item{interval}{numeric the serial number of genotype you want to calculate eg.c(1:10) or c(1,5,9,12,56)}

\item{cutp}{double, cutlevel you want, default is 0.05(P<=0.05)}
}
\value{
list(cut1,cut2), cut1 is calculate by sort(max(LR for single permutation))cutp;cut2 is calculate by max(sort(LR for single permutation)cutp)
}
\description{
hfun_permutation do permutation test
}
